audio/pyproject.toml
[project]
name = "audio-ai-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "gtts>=2.5.4",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "numpy>=2.4.4",
    "openai>=2.31.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
